fix(android): Manually capture an initial session if autocapture enabled
When configuring via unity GameObject, the script is initialized _after_ the first
Activity is loaded, so the initial session is not captured.

When configuring via code in Unity, if the code is in a behavior, then
the script is also initialized after the first activity is loaded, so
the initial session is not captured.

Forcing the initial session to be captured handles both cases.