fix: Allow setting endpoints during initialization
An event could occur between when Bugsnag is initialized and the
endpoints are set for an on-premise installation. This change surfaces
endpoint configuration during initialization and exposes input fields in
the GUI for the endpoints. If empty, the custom fields are ignored,
otherwise they replace the values of the notify and session endpoints
respectively.